Add files via upload
It should be noted here that although some groups utilized their access to the transit data for secondary eclipse signatures, the data which we analysed did not have clear signature of secondary eclipse. Although some hint of the eclipse could be seen at phase -0.4 in the final figure, further analysis was out of scope due to weak signal.